requirements.txt
flask
python-dotenv
flask-wtf
flask-security
email_validator
passlib